% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_compression.R
\name{catalogCompression}
\alias{catalogCompression}
\title{Compress the las files inside a folder}
\usage{
catalogCompression(ctg_folder, overwrite = FALSE, remove_original = FALSE)
}
\arguments{
\item{ctg_folder}{Path pointing to a folder filled with las files.}

\item{overwrite}{If set to \code{TRUE}, overwrite the existing laz files.}

\item{remove_original}{If \code{TRUE}, remove the original las files.}
}
\description{
Compress the las files inside a folder
}
\note{
One important limitation: function requires to open the LAS file. With lidR,
the file is read into R and then written back as LAZ. For very large LAS
files, laszip from lastools (rapidLasso) would be a more efficient solution.
}
