% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_stats.R
\name{.global_stats_ptype}
\alias{.global_stats_ptype}
\title{Prototypes for all the files containing the computed stats}
\usage{
.global_stats_ptype
}
\description{
Prototypes for all the files containing the computed stats
}
