% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_classification.R
\name{groundClassification}
\alias{groundClassification}
\title{Perform a ground classification}
\usage{
groundClassification(
  ctg_folder,
  overwrite = FALSE,
  n_workers = NULL,
  parallel = FALSE
)
}
\arguments{
\item{ctg_folder}{Folder to open as \code{lasCatalog}.}

\item{overwrite}{If the output folder already exists, skip the computation
by default. Otherwise, set this to \code{TRUE}.}

\item{n_workers}{When parallel computation is activated, the number of worker
or CPU cores to use.}

\item{parallel}{If \code{TRUE}, compute the new tiles in asynchronous way.}
}
\description{
Classify ground points using Progressive TIN Densification (PTD) algorithm.
}
