% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_catalog.R
\name{openCatalog}
\alias{openCatalog}
\title{Open a LAScatalog with project common options}
\usage{
openCatalog(ctg_path, check = FALSE)
}
\arguments{
\item{ctg_path}{Folder containing all the laz/las files.}

\item{check}{If \code{TRUE}, display a \code{\link[lidR:las_check]{lidR::las_check()}} message.}
}
\description{
Open a LAScatalog with project common options
}
