% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_spatial.R
\name{openSpatial}
\alias{openSpatial}
\title{Read a spatial vector layer and align its CRS with a LAS catalog}
\usage{
openSpatial(vector_path, ctg, layer_name = NULL)
}
\arguments{
\item{vector_path}{Path pointing to a file readable by \code{\link[sf:st_read]{sf::st_read()}}.}

\item{ctg}{Optional. A \code{lidR::LAScatalog} to match the CRS.}

\item{layer_name}{Optional. If it is provided, the function will try to load
the layer with this name.}
}
\value{
An \code{sf} object (AOI) in the catalog CRS (if both CRSs are defined).
}
\description{
Reads a vector dataset that could contain
\itemize{
\item The area of interest to perform regional statistics
\item Points with ground reference data to compute height RMSE.
}

After reading the file, the function transforms it to match the CRS of the
provided \code{LAScatalog} when necessary.
}
