% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_stats.R
\name{openStatsFolder}
\alias{openStatsFolder}
\title{Create the folder to store the stats from the current mission}
\usage{
openStatsFolder(project_dir, area_name, time, height)
}
\arguments{
\item{project_dir}{The directory containing the project files.}

\item{area_name}{The name of the flown area.}

\item{time}{The moment in which the flight was conducted (\code{night} or \code{day}).}

\item{height}{The flight height used to collect the point cloud.}
}
\description{
Create the folder to store the stats from the current mission
}
