% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{read_stats_csv}
\alias{read_stats_csv}
\title{Read a statistics CSV from a folder}
\usage{
read_stats_csv(folder_path, stat_prefix)
}
\arguments{
\item{folder_path}{Character scalar. Path to a folder expected to contain
exactly one CSV file.}

\item{stat_prefix}{The character corresponding with the required stat file
to find.}
}
\value{
A tibble containing the statistics read from the CSV file.
}
\description{
Reads the required CSV stats file contained inside a statistics subfolder.
}
\details{
The function expects exactly one \code{.csv} file in \code{folder_path}. If no CSV file
or more than one CSV file is found, an error is raised.
}
