% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retile_catalog.R
\name{retileCatalog}
\alias{retileCatalog}
\title{Retile the original point cloud files}
\usage{
retileCatalog(ctg_path, chunk_opts, n_workers = NULL, parallel = TRUE)
}
\arguments{
\item{ctg_path}{Folder containing the las/laz files to create the catalog.}

\item{chunk_opts}{A list with the required parameters to retile the
\code{LAScatalog}. It comes from the function \code{checkTileGrid}.}

\item{n_workers}{When parallel computation is activated, the number of worker
or CPU cores to use.}

\item{parallel}{If \code{TRUE}, compute the new tiles in asynchronous way.}
}
\value{
The output directory where the new tiles are stored.
}
\description{
In this project, the point clouds are either highly overlapped or in a unique
file. The retailing process is mandatory to efficiently process the data.
}
\note{
The output directory is set automatically inside the catalog folder.
}
