% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_stats.R
\name{targetStats}
\alias{targetStats}
\title{Compute the RMSE and associated statistics from ground reference targets}
\usage{
targetStats(
  ctg_path,
  pnts_path,
  truth_col,
  class_col,
  layer_name = NULL,
  buffer_m = 0.5,
  incoh_thresh = 0.2,
  out_folder = "",
  id_col = NULL,
  ground_filter = TRUE,
  convert_heights = FALSE,
  is_ellip = FALSE,
  geoid_height_col = NULL
)
}
\arguments{
\item{ctg_path}{Folder containing the files to create a \code{lidR::LAScatalog}.}

\item{pnts_path}{Path to the spatial file containing the target points.}

\item{truth_col}{Name of numeric column holding ground-truth height.}

\item{class_col}{The name of the column holding the available target classes.}

\item{layer_name}{Layer name; if \code{NULL}, the first layer is used.}

\item{buffer_m}{Buffer radius in meters to apply to each target.}

\item{incoh_thresh}{Absolute residual threshold for incoherency.}

\item{out_folder}{An optional folder to save the files with the stats.}

\item{id_col}{Optional ID column; default uses row index.}

\item{ground_filter}{If \code{TRUE}, keep only points classified as ground (2).}

\item{convert_heights}{If \code{TRUE}, convert target heights between ellipsoidal
and orthometric heights.}

\item{is_ellip}{Are input target heights in ellipsoidal system?}

\item{geoid_height_col}{Column name inside the \code{sf} object with the height
of the geoid, used to compute ellipsoidal heights.}
}
\value{
Invisibly returns \code{NULL}. Writes one GPKG and one CSV file per target class.
}
\description{
The stats are computed based on the target class, which could be one of the
followings:
}
\details{
\itemize{
\item \code{road} - Points located in roads or good paths with no vegetation cover.
\item \code{shrub} - Areas covered with bushes and shrubs with no or sparse trees.
\item \code{open_treed} - Areas with sparse trees but with understory.
\item \code{dense_treed} - High dense vegetated areas.
}
}
