% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retile_catalog.R
\name{checkTileGrid}
\alias{checkTileGrid}
\title{Inspect and decide the best chunk parameters to retile the catalog}
\usage{
checkTileGrid(
  ctg_folder,
  out_dir,
  chunk_size = 250,
  chunk_buffer = 0,
  alignment = TRUE,
  check = FALSE
)
}
\arguments{
\item{ctg_folder}{Directory where the point clouds are stored.}

\item{out_dir}{Path to the \code{results} directory.}

\item{chunk_size}{The desired \code{opt_chunk_size}.}

\item{chunk_buffer}{The desired \code{opt_buffer_size}.}

\item{alignment}{If \code{TRUE}, align the grid with the catalog bounds.}

\item{check}{If \code{TRUE}, produce a fast check of the LAScatalog.}
}
\description{
The grid params are stored inside the \code{results} to open later inside the
function \code{\link[=retileCatalog]{retileCatalog()}}.
}
\details{
The produced grid is saved inside the \code{results} directory as a png plot.
}
