% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_stats.R
\name{compute_target_metrics}
\alias{compute_target_metrics}
\title{Compute metrics from one point cloud within one ground reference point}
\usage{
compute_target_metrics(las, pnt_id, pnt_height, incoh_thresh)
}
\arguments{
\item{las}{A clipped LAS object.}

\item{pnt_id}{The ID of the current point to compute the stats.}

\item{pnt_height}{Ground reference height.}

\item{incoh_thresh}{Heigh value acting as a threshold to determine invalid
heigh values.}
}
\description{
This function is called after \code{\link[lidR:clip_roi]{lidR::clip_roi()}}, obtaining a list of point
clouds for each ground reference point. Call this function inside \code{\link[=Map]{Map()}}.
}
\details{
The derived metrics are (see \code{.global_stats_ptype} prototype):
\itemize{
\item n_pnts = Total number of points inside the buffer area.
\item n_pnts_last = Number of points being last returns.
\item n_pnts_last_incoherent = Last returns where the error between the point
height and the ground reference value of the target is > \code{incoh_thresh}.
\item intensity_avg = The mean intensity between all the valid points.
\item thickness = Difference between max and min height among last returns.
\item sum_errs_last = The sum of the height errors from last return points.
\item sum_errs_last_sq = The sum of squared height errors from last return points.
}
}
