% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_stats.R
\docType{data}
\name{.global_stats_ptype}
\alias{.global_stats_ptype}
\title{Prototypes for all the files containing the computed stats}
\format{
An object of class \code{list} of length 14.
}
\usage{
.global_stats_ptype
}
\description{
Prototypes for all the files containing the computed stats
}
\keyword{datasets}
